requirements/local.txt
# Development dependencies
-r base.txt
ipdb==0.13.13
django-debug-toolbar==4.3.0
django-extensions==3.2.3
pytest==8.0.0
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
coverage==7.3.2
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.7.1
django-stubs==4.2.7
django-stubs-ext==4.2.7
types-requests==2.31.0.10
django-formtools>=2.4
